Add option to avoid killing the extractor
Adds a command line flag `no_idle_exit` to avoid having the extractor process to
finish in case it doesn't detect messages for a while. Keeps the default
behavior, so this option is opt-in.

When developing and testing, it's useful to manually generate traffic and
inspect the components of the monitoring pipeline. Since this is a slow process,
the extractor keeps finishing.`