add: initial p2p-extractor
The P2P extractor starts a TCP server and listens for connections
from a Bitcoin node. A node can be configured to connect to it with
e.g. --addnode=127.0.0.1:9333. Once connected, the extractor and the
node will do a version handshake. Once that's done, the extractor can
perform measurements over the P2P interface of the node. For example,
it can send pings and record the time it takes for a pong response to
arrive. This gives insights about the nodes processing backlog.